refactor(auth): aplica tipagem forte e remove any do módulo de autenticação
